test(server): put the pool-cost measurement under the module it justifies
A _test.go here exists only under the name of a module it tests, and there is
no poolcost.go. The measurement belongs to readiness.go anyway: it is the
evidence that deferring registration is worth a gate, so it should sit beside
the gate rather than in a file named after the number it produces.
